web/requirements-web.txt
fastapi
uvicorn[standard]
python-dotenv
jinja2
